# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Red Front Hiring Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  owner: "owner",
  manager: "manager",
} as const;

export interface User {
  id: number;
  email: string;
  role: UserRole;
}

export interface LoginResponse {
  message: string;
  user: User;
}

export type ApplicantRawPayload = { [key: string]: unknown };

export type ApplicantStatus =
  (typeof ApplicantStatus)[keyof typeof ApplicantStatus];

export const ApplicantStatus = {
  new: "new",
  reviewed: "reviewed",
  interviewing: "interviewing",
  hired: "hired",
  rejected: "rejected",
  archived: "archived",
} as const;

export interface Applicant {
  id: number;
  location?: string;
  position?: string;
  name: string;
  email?: string;
  address?: string;
  city?: string;
  state?: string;
  zip?: string;
  phoneHome?: string;
  phoneBusiness?: string;
  phoneCell?: string;
  dateCanStart?: string;
  salaryDesired?: string;
  hasHighSchoolDiploma?: boolean;
  rawPayload?: ApplicantRawPayload;
  status: ApplicantStatus;
  createdAt: string;
  updatedAt: string;
  notesCount?: number;
}

export type UpdateApplicantRequestStatus =
  (typeof UpdateApplicantRequestStatus)[keyof typeof UpdateApplicantRequestStatus];

export const UpdateApplicantRequestStatus = {
  new: "new",
  reviewed: "reviewed",
  interviewing: "interviewing",
  hired: "hired",
  rejected: "rejected",
  archived: "archived",
} as const;

export interface UpdateApplicantRequest {
  status?: UpdateApplicantRequestStatus;
}

export interface Note {
  id: number;
  applicantId: number;
  body: string;
  author: string;
  createdAt: string;
}

export interface CreateNoteRequest {
  body: string;
}

export interface Cf7WebhookPayload {
  [key: string]: unknown;
}

export interface DashboardStats {
  total: number;
  new: number;
  reviewed: number;
  interviewing: number;
  hired: number;
  rejected: number;
  archived: number;
  recentApplicants: Applicant[];
}

export type ListApplicantsParams = {
  status?: ListApplicantsStatus;
  search?: string;
  location?: string;
  position?: string;
};

export type ListApplicantsStatus =
  (typeof ListApplicantsStatus)[keyof typeof ListApplicantsStatus];

export const ListApplicantsStatus = {
  new: "new",
  reviewed: "reviewed",
  interviewing: "interviewing",
  hired: "hired",
  rejected: "rejected",
  archived: "archived",
} as const;