# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Red Front Hiring Manager API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  message: zod.string(),
  user: zod.object({
    id: zod.number(),
    email: zod.string(),
    role: zod.enum(["owner", "manager"]),
  }),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  email: zod.string(),
  role: zod.enum(["owner", "manager"]),
});

/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary List all applicants
 */
export const ListApplicantsQueryParams = zod.object({
  status: zod
    .enum(["new", "reviewed", "interviewing", "hired", "rejected", "archived"])
    .optional(),
  search: zod.coerce.string().optional(),
  location: zod.coerce.string().optional(),
  position: zod.coerce.string().optional(),
});

export const ListApplicantsResponseItem = zod.object({
  id: zod.number(),
  location: zod.string().optional(),
  position: zod.string().optional(),
  name: zod.string(),
  email: zod.string().optional(),
  address: zod.string().optional(),
  city: zod.string().optional(),
  state: zod.string().optional(),
  zip: zod.string().optional(),
  phoneHome: zod.string().optional(),
  phoneBusiness: zod.string().optional(),
  phoneCell: zod.string().optional(),
  dateCanStart: zod.string().optional(),
  salaryDesired: zod.string().optional(),
  hasHighSchoolDiploma: zod.boolean().optional(),
  rawPayload: zod.object({}).passthrough().optional(),
  status: zod.enum([
    "new",
    "reviewed",
    "interviewing",
    "hired",
    "rejected",
    "archived",
  ]),
  createdAt: zod.date(),
  updatedAt: zod.date(),
  notesCount: zod.number().optional(),
});
export const ListApplicantsResponse = zod.array(ListApplicantsResponseItem);

/**
 * @summary Get applicant detail
 */
export const GetApplicantParams = zod.object({
  id: zod.coerce.number(),
});

export const GetApplicantResponse = zod.object({
  id: zod.number(),
  location: zod.string().optional(),
  position: zod.string().optional(),
  name: zod.string(),
  email: zod.string().optional(),
  address: zod.string().optional(),
  city: zod.string().optional(),
  state: zod.string().optional(),
  zip: zod.string().optional(),
  phoneHome: zod.string().optional(),
  phoneBusiness: zod.string().optional(),
  phoneCell: zod.string().optional(),
  dateCanStart: zod.string().optional(),
  salaryDesired: zod.string().optional(),
  hasHighSchoolDiploma: zod.boolean().optional(),
  rawPayload: zod.object({}).passthrough().optional(),
  status: zod.enum([
    "new",
    "reviewed",
    "interviewing",
    "hired",
    "rejected",
    "archived",
  ]),
  createdAt: zod.date(),
  updatedAt: zod.date(),
  notesCount: zod.number().optional(),
});

/**
 * @summary Update applicant status
 */
export const UpdateApplicantParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateApplicantBody = zod.object({
  status: zod
    .enum(["new", "reviewed", "interviewing", "hired", "rejected", "archived"])
    .optional(),
});

export const UpdateApplicantResponse = zod.object({
  id: zod.number(),
  location: zod.string().optional(),
  position: zod.string().optional(),
  name: zod.string(),
  email: zod.string().optional(),
  address: zod.string().optional(),
  city: zod.string().optional(),
  state: zod.string().optional(),
  zip: zod.string().optional(),
  phoneHome: zod.string().optional(),
  phoneBusiness: zod.string().optional(),
  phoneCell: zod.string().optional(),
  dateCanStart: zod.string().optional(),
  salaryDesired: zod.string().optional(),
  hasHighSchoolDiploma: zod.boolean().optional(),
  rawPayload: zod.object({}).passthrough().optional(),
  status: zod.enum([
    "new",
    "reviewed",
    "interviewing",
    "hired",
    "rejected",
    "archived",
  ]),
  createdAt: zod.date(),
  updatedAt: zod.date(),
  notesCount: zod.number().optional(),
});

/**
 * @summary Delete applicant
 */
export const DeleteApplicantParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteApplicantResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary List notes for applicant
 */
export const ListNotesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListNotesResponseItem = zod.object({
  id: zod.number(),
  applicantId: zod.number(),
  body: zod.string(),
  author: zod.string(),
  createdAt: zod.date(),
});
export const ListNotesResponse = zod.array(ListNotesResponseItem);

/**
 * @summary Add note to applicant
 */
export const CreateNoteParams = zod.object({
  id: zod.coerce.number(),
});

export const CreateNoteBody = zod.object({
  body: zod.string(),
});

/**
 * @summary WordPress Contact Form 7 webhook
 */
export const WebhookCf7Body = zod.record(zod.string(), zod.unknown());

export const WebhookCf7Response = zod.object({
  message: zod.string(),
});

/**
 * @summary Get dashboard stats
 */
export const GetStatsResponse = zod.object({
  total: zod.number(),
  new: zod.number(),
  reviewed: zod.number(),
  interviewing: zod.number(),
  hired: zod.number(),
  rejected: zod.number(),
  archived: zod.number(),
  recentApplicants: zod.array(
    zod.object({
      id: zod.number(),
      location: zod.string().optional(),
      position: zod.string().optional(),
      name: zod.string(),
      email: zod.string().optional(),
      address: zod.string().optional(),
      city: zod.string().optional(),
      state: zod.string().optional(),
      zip: zod.string().optional(),
      phoneHome: zod.string().optional(),
      phoneBusiness: zod.string().optional(),
      phoneCell: zod.string().optional(),
      dateCanStart: zod.string().optional(),
      salaryDesired: zod.string().optional(),
      hasHighSchoolDiploma: zod.boolean().optional(),
      rawPayload: zod.object({}).passthrough().optional(),
      status: zod.enum([
        "new",
        "reviewed",
        "interviewing",
        "hired",
        "rejected",
        "archived",
      ]),
      createdAt: zod.date(),
      updatedAt: zod.date(),
      notesCount: zod.number().optional(),
    }),
  ),
});