# lib/api-zod/src/generated/types/applicant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Red Front Hiring Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicantRawPayload } from "./applicantRawPayload";
import type { ApplicantStatus } from "./applicantStatus";

export interface Applicant {
  id: number;
  location?: string;
  position?: string;
  name: string;
  email?: string;
  address?: string;
  city?: string;
  state?: string;
  zip?: string;
  phoneHome?: string;
  phoneBusiness?: string;
  phoneCell?: string;
  dateCanStart?: string;
  salaryDesired?: string;
  hasHighSchoolDiploma?: boolean;
  rawPayload?: ApplicantRawPayload;
  status: ApplicantStatus;
  createdAt: Date;
  updatedAt: Date;
  notesCount?: number;
}