# lib/api-zod/src/generated/types/cf7WebhookPayload.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Red Front Hiring Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface Cf7WebhookPayload {
  [key: string]: unknown;
}