# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Red Front Hiring Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { Applicant } from "./applicant";

export interface DashboardStats {
  total: number;
  new: number;
  reviewed: number;
  interviewing: number;
  hired: number;
  rejected: number;
  archived: number;
  recentApplicants: Applicant[];
}