# lib/api-zod/src/generated/types/updateApplicantRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Red Front Hiring Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateApplicantRequestStatus } from "./updateApplicantRequestStatus";

export interface UpdateApplicantRequest {
  status?: UpdateApplicantRequestStatus;
}